Added in handling for new_users (name, prof and group size),  fixed up eventhandler and fixed subjects and ids for mydicts
