+++ mod/mydicts.py
@@ -13,13 +13,23 @@ roles = {
     "NEWBIE": 762190942316134400,
     "MUTED": 701972373053636649
 }
+
 group_roles = {
-    "DUO": 765011328532086805,
-    "TRIO": 765011376976166942,
-    "QUARTET": 765011399524614169,
-    "QUINTET": 765011419598815233,
-    "DECUPLET": 765011478775463947,
-    "VIGINTET": 765011552007880736
+    "🕑": "duo",
+    "🕒": "trio",
+    "🕓": "quartet",
+    "🕔": "qunitet",
+    "🕙": "decuplet",
+    "👪": "vigintet"
+}
+
+group_roles_ids = {
+    "duo": 765011328532086805,
+    "trio": 765011376976166942,
+    "quartet": 765011399524614169,
+    "qunitet": 765011419598815233,
+    "decuplet": 765011478775463947,
+    "vigintet": 765011552007880736,
 }
 
 channels = {
@@ -29,85 +39,86 @@ channels = {
     "BULK_DELETES": 786003742813192233,
     "JOIN_LEAVES": 786016910668070952,
     "ERROR_ROOM": 828638567236108308,
-    "INTRO_CHANNEL": 830914205078782072,
+    "PERSONALIZE_CHANNEL": 830914205078782072,
     "WELCOME_CHANNEL": 700214669003980801,
     "MEMBER_COUNT_CHANNEL": 764418047246729227,
     "MEMBERS_IN_VC_COUNT_CHANNEL": 764427421876748298
 }
 
-csec_subjects = {
-"📈":"csec add maths",
-"🌾":"csec agricultural science",
-"🧬":"csec biology",
-"🏗":"csec building tech",
-"🌋":"csec caribbean history",
-"🧪":"csec certificate in business studies",
-"🏛":"csec chemistry",
-"🧵":"csec clothing and textiles",
-"💳":"csec economics",
-"⌨":"csec edpm",
-"🔌":"csec electrical engineering",
-"📕":"csec english a",
-"📗":"csec english b",
-"🍽":"csec food and nutrition",
-"🥖":"csec french",
-"🗺":"csec geography",
-"🧠":"csec hsb",
-"🖥":"csec information technology",
-"⚗":"csec integrated science",
-"📉":"csec maths",
-"⚙":"csec mechanical engineering",
-"🎶":"csec music",
-"⚽":"csec office admin",
-"📠":"csec physical education",
-"🧲":"csec physics",
-"💷":"csec poa",
-"💵":"csec pob",
-"🙏":"csec religious education",
-"🏡":"csec resources and family management",
-"🎎":"csec social studies",
-"💃":"csec spanish",
-"✏":"csec technical drawing",
-"🎭":"csec theatre arts",
-"🎨":"csec visual arts"
-}
-
-cape_subjects = {
-"💲":"cape accounting",
-"🥬":"cape agricultural science",
-"🎮":"cape animation and game design",
-"📊":"cape applied maths",
-"🌌":"cape arts and design",
-"🦠":"cape biology",
-"🛠":"cape building and mechanical engineering drawing",
-"📙":"cape caribbean studies",
-"🌡":"cape chemistry",
-"📘":"cape communication studies",
-"💻":"cape computer science",
-"🎥":"cape digital media",
-"💱":"cape economics",
-"💡":"cape electrical electronic engineering tech",
-"💰":"cape entrepreneurship",
-"🏞":"cape environmental science",
-"🗂":"cape financial services studies",
-"🍔":"cape food and nutrition",
-"🥖":"cape french",
-"🌍":"cape geography",
-"♻":"cape green engineering",
-"🗾":"cape history",
-"💾":"cape information tech",
-"🔢":"cape integrated maths",
-"⚖":"cape law",
-"📖":"cape literature in english",
-"🏧":"cape logistics and supplies chain oppositions",
-"💼":"cape management of business",
-"💃":"cape performing arts",
-"🏀":"cape physical education",
-"🔋":"cape physics",
-"📐":"cape pure maths",
-"🗿":"cape sociology",
-"💃":"cape spanish",
-"🏖":"cape tourism"
+subjects= {
+    "CSEC": {
+        "📈":"csec add maths",
+        "🌾":"csec agricultural science",
+        "🧬":"csec biology",
+        "🏗":"csec building tech",
+        "🌋":"csec caribbean history",
+        "🧪":"csec certificate in business studies",
+        "🏛":"csec chemistry",
+        "🧵":"csec clothing and textiles",
+        "💳":"csec economics",
+        "⌨":"csec edpm",
+        "🔌":"csec electrical engineering",
+        "📕":"csec english a",
+        "📗":"csec english b",
+        "🍽":"csec food and nutrition",
+        "🥖":"csec french",
+        "🗺":"csec geography",
+        "🧠":"csec hsb",
+        "🖥":"csec information technology",
+        "⚗":"csec integrated science",
+        "📉":"csec maths",
+        "⚙":"csec mechanical engineering",
+        "🎶":"csec music",
+        "⚽":"csec office admin",
+        "📠":"csec physical education",
+        "🧲":"csec physics",
+        "💷":"csec poa",
+        "💵":"csec pob",
+        "🙏":"csec religious education",
+        "🏡":"csec resources and family management",
+        "🎎":"csec social studies",
+        "💃":"csec spanish",
+        "✏":"csec technical drawing",
+        "🎭":"csec theatre arts",
+        "🎨":"csec visual arts"
+    },
+    "CAPE": {
+        "💲":"cape accounting",
+        "🥬":"cape agricultural science",
+        "🎮":"cape animation and game design",
+        "📊":"cape applied maths",
+        "🌌":"cape arts and design",
+        "🦠":"cape biology",
+        "🛠":"cape building and mechanical engineering drawing",
+        "📙":"cape caribbean studies",
+        "🌡":"cape chemistry",
+        "📘":"cape communication studies",
+        "💻":"cape computer science",
+        "🎥":"cape digital media",
+        "💱":"cape economics",
+        "💡":"cape electrical electronic engineering tech",
+        "💰":"cape entrepreneurship",
+        "🏞":"cape environmental science",
+        "🗂":"cape financial services studies",
+        "🍔":"cape food and nutrition",
+        "🥖":"cape french",
+        "🌍":"cape geography",
+        "♻":"cape green engineering",
+        "🗾":"cape history",
+        "💾":"cape information tech",
+        "🔢":"cape integrated maths",
+        "⚖":"cape law",
+        "📖":"cape literature in english",
+        "🏧":"cape logistics and supplies chain oppositions",
+        "💼":"cape management of business",
+        "💃":"cape performing arts",
+        "🏀":"cape physical education",
+        "🔋":"cape physics",
+        "📐":"cape pure maths",
+        "🗿":"cape sociology",
+        "💃":"cape spanish",
+        "🏖":"cape tourism"
+    }
 }
 
 logs = []
